nit: rename wlm process-collector to process-language-collector
