+++ .golangci.yml
@@ -464,7 +464,7 @@ issues:
     - path: comp/core/workloadmeta/collectors/internal/podman/podman.go
       linters:
         - pkgconfigusage
-    - path: comp/core/workloadmeta/collectors/internal/process/process_collector.go
+    - path: comp/core/workloadmeta/collectors/internal/processlanguage/process_collector.go
       linters:
         - pkgconfigusage
     - path: comp/core/workloadmeta/collectors/util/process_util_linux.go
