+++ comp/core/workloadmeta/collectors/catalog-core/options.go
@@ -21,7 +21,7 @@ import (
 	"github.com/DataDog/datadog-agent/comp/core/workloadmeta/collectors/internal/kubemetadata"
 	"github.com/DataDog/datadog-agent/comp/core/workloadmeta/collectors/internal/nvml"
 	"github.com/DataDog/datadog-agent/comp/core/workloadmeta/collectors/internal/podman"
-	"github.com/DataDog/datadog-agent/comp/core/workloadmeta/collectors/internal/process"
+	"github.com/DataDog/datadog-agent/comp/core/workloadmeta/collectors/internal/processlanguage"
 	remoteprocesscollector "github.com/DataDog/datadog-agent/comp/core/workloadmeta/collectors/internal/remote/processcollector"
 )
 
@@ -38,7 +38,7 @@ func getCollectorOptions() []fx.Option {
 		kubemetadata.GetFxOptions(),
 		podman.GetFxOptions(),
 		remoteprocesscollector.GetFxOptions(),
-		process.GetFxOptions(),
+		processlanguage.GetFxOptions(),
 		nvml.GetFxOptions(),
 	}
 }
