+++ comp/core/workloadmeta/collectors/internal/processlanguage/process_collector.go
@@ -5,8 +5,8 @@
 
 //go:build linux
 
-// Package process implements the local process collector for Workloadmeta.
-package process
+// Package process implements the process language collector for Workloadmeta.
+package processlanguage
 
 import (
 	"context"
@@ -27,8 +27,8 @@ import (
 )
 
 const (
-	collectorID       = "local-process-collector"
-	componentName     = "workloadmeta-process"
+	collectorID       = "process-language-collector"
+	componentName     = "workloadmeta-process-language"
 	cacheValidityNoRT = 2 * time.Second
 )
 
@@ -47,9 +47,9 @@ type collector struct {
 	containerProvider proccontainers.ContainerProvider
 }
 
-// NewCollector returns a new local process collector provider and an error.
+// NewProcessLanguageCollector returns a new local process collector provider and an error.
 // Currently, this is only used on Linux when language detection and run in core agent are enabled.
-func NewCollector() (workloadmeta.CollectorProvider, error) {
+func NewProcessLanguageCollector() (workloadmeta.CollectorProvider, error) {
 	wlmExtractor := processwlm.GetSharedWorkloadMetaExtractor(pkgconfigsetup.SystemProbe())
 	processData := NewProcessData()
 	processData.Register(wlmExtractor)
@@ -69,7 +69,7 @@ func NewCollector() (workloadmeta.CollectorProvider, error) {
 
 // GetFxOptions returns the FX framework options for the collector
 func GetFxOptions() fx.Option {
-	return fx.Provide(NewCollector)
+	return fx.Provide(NewProcessLanguageCollector)
 }
 
 func (c *collector) Start(ctx context.Context, store workloadmeta.Component) error {
