+++ comp/core/workloadmeta/collectors/internal/processlanguage/process_collector_test.go
@@ -7,7 +7,7 @@
 
 // Package process implements the local process collector for
 // Workloadmeta.
-package process
+package processlanguage
 
 import (
 	"context"
