+++ comp/core/workloadmeta/collectors/internal/processlanguage/process_data.go
@@ -6,7 +6,7 @@
 //go:build linux
 
 // Package process implements the local process collector for Workloadmeta.
-package process
+package processlanguage
 
 import (
 	"time"
